requirements.txt
python-telegram-bot==20.8
python-dotenv==1.0.0
requests==2.31.0
aiohttp>=3.8.0
psutil>=5.9.0